python/requirements.txt
python-multipart
fastapi
uvicorn[standard]
celery[redis]
redis
transformers
torch
opencv-python-headless
open3d==0.19.0
trimesh
numpy
python-dotenv